Moved toJson methods
